+++ src/services/vehiculos.service.js
@@ -1,19 +1,16 @@
 const vehiculosRepository = require('../repositories/vehiculos.repository');
 
 const getAll = async () => {
-    return vehiculosRepository.findAll();
+    const result = await vehiculosRepository.findAll();
+    return result.length > 0 ? { statusCode: 200, message: 'Vehiculos encontrados', success: true, data: result } : { statusCode: 404, message: 'No se encontraron vehiculos', success: false, data: [] };
 };
 
 const getById = async (id) => {
     const record = await vehiculosRepository.findById(id);
 
-    if (!record) {
-        const error = new Error('vehiculos no encontrado');
-        error.statusCode = 404;
-        throw error;
-    }
+    const result = record ? { statusCode: 200, message: 'Vehiculo encontrado', success: true, data: record } : { statusCode: 404, message: 'Vehiculo no encontrado', success: false, data: [] };
 
-    return record;
+    return result;
 };
 
 const create = async (data) => {
@@ -21,15 +18,29 @@ const create = async (data) => {
 };
 
 const update = async (id, data) => {
-    await getById(id);
+    const result = await getById(id);
+
+    console.log('Result from getById:', result); // Debugging line.
+    if ( result.statusCode === 404) {
+        const error = 'No se encontró el vehiculo, No se puede actualizar';
+        return { statusCode: 404, message: error, success: false, data: [] };
+    } 
 
     return vehiculosRepository.update(id, data);
+
 };
 
 const remove = async (id) => {
-    await getById(id);
+    const result = await getById(id);
 
-    return vehiculosRepository.remove(id);
+    if (result.statusCode === 404) {
+        const error = 'No se encontró el vehiculo, No se puede eliminar';
+        return { statusCode: 404, message: error, success: false, data: [] };
+    } 
+
+    const data = await vehiculosRepository.remove(id)
+    
+    return  { statusCode: 200, message: 'Vehiculo eliminado correctamente', success: true, data };
 };
 
 module.exports = {
